sneaker-drops-api-4 repository.

+++ src/main/java/com/pluralsight/sneakerdrops/controller/SneakerController.java
@@ -2,9 +2,12 @@ package com.pluralsight.sneakerdrops.controller;
 
 import com.pluralsight.sneakerdrops.models.Sneaker;
 import com.pluralsight.sneakerdrops.service.SneakerService;
+import jakarta.validation.Valid;
+import org.hibernate.annotations.NotFound;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
+import org.springframework.web.server.ResponseStatusException;
 
 import java.util.List;
 
@@ -31,24 +34,34 @@ public class SneakerController {
 
     @GetMapping("/{id}")
     public Sneaker getById(@PathVariable long id) {
-        return sneakerService.byId(id);
+        Sneaker sneaker = sneakerService.byId(id);
+        if (sneaker == null) {
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "No Sneaker with id " + id);
+        }
+        return sneaker;
     }
 
     @PostMapping
-    public ResponseEntity<Sneaker> create(@RequestBody Sneaker sneaker) {
+    public ResponseEntity<Sneaker> create(@Valid @RequestBody Sneaker sneaker) {
         Sneaker saved = sneakerService.createSneaker(sneaker);
         return ResponseEntity.status(HttpStatus.CREATED).body(saved);
     }
 
     @PutMapping("/{id}")
-    public Sneaker update(@PathVariable long id, @RequestBody Sneaker sneaker) {
-        return sneakerService.updateSneaker(id, sneaker);
+    public Sneaker update(@PathVariable long id,@Valid @RequestBody Sneaker sneaker) {
+        Sneaker saved =  sneakerService.updateSneaker(id, sneaker);
+        if(saved == null) {
+            throw  new ResponseStatusException(HttpStatus.NOT_FOUND, "No sneaker with id " + id);
+        }
+        return saved;
     }
 
     @DeleteMapping("/{id}")
-    public ResponseEntity<Void> delete(@PathVariable long id){
-     sneakerService.deleteSneaker(id);
-     return ResponseEntity.noContent().build();
+    public ResponseEntity<Void> delete(@PathVariable long id) {
+        if (sneakerService.byId(id) == null) {
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "No sneaker with id " + id);
+        }
+        sneakerService.deleteSneaker(id);
+        return ResponseEntity.noContent().build();
     }
-
 }
